+++ src/components/App.tsx
@@ -1,4 +1,5 @@
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
+import { Offer } from '../types';
 import MainPage from './Main';
 import LoginPage from './Login';
 import FavoritesPage from './Favorites';
@@ -6,13 +7,13 @@ import OfferPage from './Offer';
 import Page404 from './404';
 import PrivateRoute from './PrivateRoute';
 
-function App({ numberOfOffers }: { numberOfOffers: number }) {
+function App({ offers }: { offers: Offer[] }) {
   const isAuthenticated = false;
 
   return (
     <BrowserRouter>
       <Routes>
-        <Route path="/" element={<MainPage numberOfOffers={numberOfOffers} />} />
+        <Route path="/" element={<MainPage offers={offers}/>} />
         <Route path="/login" element={<LoginPage />} />
         <Route
           path="/favorites"
