+++ src/index.tsx
@@ -1,15 +1,14 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './components/App';
+import offers from './mocks/offers';
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
 );
 
-const numberOfOffers = 312; // Data defined here
-
 root.render(
   <React.StrictMode>
-    <App numberOfOffers={numberOfOffers} />
+    <App offers={offers} />
   </React.StrictMode>
 );
