+++ src/types.ts
@@ -1,0 +1,9 @@
+export type Offer = {
+    isPremium: boolean;
+    imageUrl: string;
+    price: number;
+    isBookmarked: boolean;
+    rating: number;
+    title: string;
+    type: string;
+};
